requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.7
django-cors==0.1
django-cors-headers==2.5.2
django-rest-framework==0.1.0
djangorestframework==3.9.1
djangorestframework-gis==0.14
djangorestframework-jwt==1.11.0
idna==2.8
mysqlclient==1.4.2.post1
Pillow==5.4.1
PyJWT==1.7.1
pytz==2018.9
requests==2.21.0
urllib3==1.24.1
